lab1/requirements.txt
gradio>=4.0.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.20.0
matplotlib>=3.7.0
requests>=2.25.0
